Add additional inputs, add test, fix annotation type, add null-safe attribute access
