+++ README.md
@@ -11,7 +11,7 @@ This is a GitHub Action to run `mypy` against your repository. It uses the new G
 Use it in your project like:
 
 (in `.github/workflows/lint.yml`)
-```
+```yaml
 name: Lint
 
 on:
@@ -24,18 +24,20 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Setup Python
-        uses: actions/setup-python@v1
+        uses: actions/setup-python@v4
         with:
-          python-version: 3.7.4
+          python-version: "3.10"
           architecture: x64
       - name: Checkout
-        uses: actions/checkout@v1
+        uses: actions/checkout@v2
       - name: Install mypy
         run: pip install mypy
       - name: Run mypy
-        uses: sasanquaneuf/mypy-github-action@releases/v1
+        uses: sasanquaneuf/mypy-github-action@releases/v2
         with:
           checkName: 'mypy'   # NOTE: this needs to be the same as the job name
+          cmd: "tox run -e type" # Optional
+          args: "--follow-imports=silent" # Optional
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 ```
